изменен тип переменных

+++ LR2git/Form1.cs
@@ -24,9 +24,9 @@ namespace LR2git
 
         private void button1_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToInt32(textBox1.Text);
+            double b = Convert.ToInt32(textBox2.Text);
+            double c;
             c = a + b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Сложение";
@@ -34,9 +34,9 @@ namespace LR2git
 
         private void button2_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToInt32(textBox1.Text);
+            double b = Convert.ToInt32(textBox2.Text);
+            double c;
             c = a - b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Вычитание";
@@ -44,9 +44,9 @@ namespace LR2git
 
         private void button3_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToInt32(textBox1.Text);
+            double b = Convert.ToInt32(textBox2.Text);
+            double c;
             c = a * b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Умножение";
@@ -54,9 +54,9 @@ namespace LR2git
 
         private void button4_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToInt32(textBox1.Text);
+            double b = Convert.ToInt32(textBox2.Text);
+            double c;
             c = a / b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Деление";
